+++ src/main/java/es/sanguino/ecommerce/infrastructure/entity/ProductEntity.java
@@ -12,12 +12,12 @@ public class ProductEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
-    private Long price;
+    private Double price;
 
     public ProductEntity() {
     }
 
-    public ProductEntity(Long id, String name, Long price) {
+    public ProductEntity(Long id, String name, Double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -39,11 +39,11 @@ public class ProductEntity {
         this.name = name;
     }
 
-    public Long getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Long price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
